# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: WizLetter API
  description: |
    WizLetter (위즈레터) backend API — AI-curated daily briefing service with coin/level gamification.

    ## Authentication
    Most endpoints require a JWT access token stored in an HttpOnly cookie (`ota_token`).
    The token is issued automatically after Kakao OAuth login. A short-lived access token
    (15 min) and a long-lived refresh token (7 days, cookie `ota_refresh`) are issued together.

    Endpoints may also accept `Authorization: Bearer <token>` as a fallback.

    ## Base URL
    - Production: `https://server.mindhacker.club`
    - Local: `http://localhost:8080`
  version: "1.0.0"
  contact:
    email: mindhacker.admin@gmail.com
servers:
  - url: https://server.mindhacker.club/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ota_token
      description: JWT access token (15 min TTL). Set automatically after Kakao login.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Fallback bearer token (same JWT as cookie).
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LevelInfo:
      type: object
      properties:
        level:
          type: integer
        total_coins:
          type: integer
        daily_limit:
          type: integer
        coin_cap:
          type: integer
        thresholds:
          type: array
          items:
            type: integer
        description:
          type: string
...
paths:
...
  /level:
    get:
      tags: [level]
      summary: Get the authenticated user's level and coin info
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Level info
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/LevelInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /level/init-earn:
    post:
      tags: [level]
      summary: Initiate coin earn for a topic view
      description: |
        Step 1 of the two-step coin earning flow. Validates all eligibility conditions
        (item exists, not already earned, run is from today, daily limit not reached)
        and records a pending entry in the cache. Call `/level/earn` after the minimum
        dwell time has elapsed.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [context_item_id]
              properties:
                context_item_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Earn initiation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: [PENDING, DUPLICATE, EXPIRED, DAILY_LIMIT]
                      required_seconds:
                        type: integer
                        description: Only present when status=PENDING
        '400':
          description: Invalid or missing context_item_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Context item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /level/earn:
    post:
      tags: [level]
      summary: Confirm coin earn after dwell time
      description: |
        Step 2 of the two-step coin earning flow. Requires a Cloudflare Turnstile token
        for bot verification. Verifies that the minimum dwell time has elapsed since
        `/level/init-earn`. Awards coins on success.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [context_item_id, turnstile_token]
              properties:
                context_item_id:
                  type: string
                  format: uuid
                turnstile_token:
                  type: string
                  description: Cloudflare Turnstile challenge token
      responses:
        '200':
          description: Earn result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      attempted:
                        type: boolean
                      earned:
                        type: boolean
                      reason:
                        type: string
                        enum: [EARNED, DUPLICATE, EXPIRED, DAILY_LIMIT, CAP_REACHED]
                      coins_earned:
                        type: integer
                      leveled_up:
                        type: boolean
                      new_level:
                        type: integer
        '400':
          description: EARN_NOT_INITIATED (earn not initiated) or TOO_EARLY (dwell time not met) or invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Bot verification failed (Turnstile)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /level/batch-earn-status:
    post:
      tags: [level]
      summary: Batch query earn status for multiple topics (max 50)
      description: |
        Returns the earn status and potential coin value for each requested topic.
        Used to render earn indicators on the topic list page.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [context_item_ids]
              properties:
                context_item_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  maxItems: 50
      responses:
        '200':
          description: Earn status per item
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        status:
                          type: string
                          enum: [PENDING, DUPLICATE, EXPIRED, DAILY_LIMIT, NOT_FOUND]
                        coins:
                          type: integer
        '400':
          description: Invalid input or too many items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Subscriptions ───────────────────────────────────────────────────────────
...
security:
  - cookieAuth: []
  - bearerAuth: []